+++ FitAppka/Service/ServiceInterface/IMeasurementsService.cs
@@ -2,9 +2,9 @@
 {
     public interface IMeasurementsService
     {
-        void AddMeasurements(short weight, int? waist);
-        bool UpdateMeasurements(int id, short weight, int? waist);
+        void AddMeasurements(double weight, int? waist);
+        bool UpdateMeasurements(int id, double weight, int? waist);
         bool DeleteMeasurement(int id);
-        double EstimateBodyFatLevel(short weight, int? waist);
+        double EstimateBodyFatLevel(double weight, int? waist);
     }
 }
